FIX: common_tools_string - module now operates on u16 string-length

+++ src/common/common_tools_string.h
@@ -91,12 +91,12 @@ void common_tools_string_trim(char* p_string);
 /*
  *
  */
-u8 common_tools_hex_string_to_byte_array(const char* hex_string, u16 hex_string_len, u8* byte_array, u16 byte_array_max_len);
+u16 common_tools_hex_string_to_byte_array(const char* hex_string, u16 hex_string_len, u8* byte_array, u16 byte_array_max_len);
 
 /*
  *
  */
-u8 common_tools_byte_array_string_to_hex_string(u8* byte_array, u8 byte_array_len, char* hex_string, u16 hex_string_max_len);
+u16 common_tools_byte_array_string_to_hex_string(u8* byte_array, u8 byte_array_len, char* hex_string, u16 hex_string_max_len);
 
 /*
  *
